update: to minecraft 1.21.1

+++ src/main/java/cn/seiua/skymatrix/event/events/ViewClipEvent.java
@@ -11,8 +11,8 @@ public class ViewClipEvent extends Event {
         this.dis = dis;
     }
 
-    public double getDis() {
-        return dis;
+    public float getDis() {
+        return (float) dis;
     }
 
     public void setDis(double dis) {
